update SidebarLayout tests
